Kill mutants.
