+++ tests/Concern/HasClassTest.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace UIAwesome\Html\Field\Tests\Concern;
+
+use UIAwesome\Html\Field\Concern\HasClass;
+
+final class HasClassTest extends \PHPUnit\Framework\TestCase
+{
+    public function testImmutability(): void
+    {
+        $instance = new class () {
+            use HasClass;
+        };
+
+        $this->assertNotSame($instance, $instance->class(''));
+    }
+}
